+++ Tests/tst_tuples.c
@@ -5,14 +5,14 @@
 
 #define EPSILON 0.00001
 
-static int	assert_float(float a, float b)
+static int	assert_float(double a, double b)
 {
-	return (fabsf(a - b) < EPSILON);
+	return (fabs(a - b) < EPSILON);
 }
 
 Test(point, expected_a_array_with_numbers_equals_params_passed) {
-	const float *tupl = point(4.3, -4.2, 3.1);
-	float expected_values[] = {4.3, -4.2, 3.1, 1.0};
+	const double *tupl = point(4.3, -4.2, 3.1);
+	double expected_values[] = {4.3, -4.2, 3.1, 1.0};
 
 	for (int i = 0; i < 4; i++) {
 		cr_assert_eq(assert_float(tupl[i], expected_values[i]), 1);
@@ -21,8 +21,8 @@ Test(point, expected_a_array_with_numbers_equals_params_passed) {
 }
 
 Test(vector, expected_a_array_with_numbers_equals_params_passed) {
-	float *tupl = point(4.3, -4.2, 3.1);
-	float expected_values[] = {4.3, -4.2, 3.1, 1.0};
+	double *tupl = point(4.3, -4.2, 3.1);
+	double expected_values[] = {4.3, -4.2, 3.1, 1.0};
 
 	for (int i = 0; i < 4; i++) {
 		cr_assert_eq(assert_float(tupl[i], expected_values[i]), 1);
@@ -31,11 +31,11 @@ Test(vector, expected_a_array_with_numbers_equals_params_passed) {
 }
 
 Test(add_vectors, expected_sum_of_vectors_a_and_b)	{
-	float *vector1 = vector(3, -2, 5);
-	float *vector2 = vector(-2, 3, 1);
-	float	expected_values[] = {1.0, 1.0, 6, 0};
+	double *vector1 = vector(3, -2, 5);
+	double *vector2 = vector(-2, 3, 1);
+	double	expected_values[] = {1.0, 1.0, 6, 0};
 
-	float *sum = add_tuples(vector1, vector2);
+	double *sum = add_tuples(vector1, vector2);
 	for (int i = 0; i < 4; i++) {
 		cr_assert_eq(assert_float(sum[i], expected_values[i]), 1);
 	}
@@ -44,11 +44,11 @@ Test(add_vectors, expected_sum_of_vectors_a_and_b)	{
 
 Test(sub_vectors, expected_sub_of_vectors_a_and_b)
 {
-	float *v1 = vector(3, 2, 1);
-	float *v2 = vector(5, 6, 7);
-	float expected_values[] = {-2, -4, -6, 0};
+	double *v1 = vector(3, 2, 1);
+	double *v2 = vector(5, 6, 7);
+	double expected_values[] = {-2, -4, -6, 0};
 
-	float *sub = sub_tuple(v1, v2);
+	double *sub = sub_tuple(v1, v2);
 	for (int i = 0; i < 4; i++) {
 		cr_assert_eq(assert_float(sub[i],expected_values[i]), 1);
 	}
@@ -58,11 +58,11 @@ Test(sub_vectors, expected_sub_of_vectors_a_and_b)
 }
 
 Test(sub_points, expected_vector) {
-	float *p1 = point(3, 2, 1);
-	float *p2 = point(5, 6, 7);
-	float expected_values[] = {-2, -4, -6, 0};
+	double *p1 = point(3, 2, 1);
+	double *p2 = point(5, 6, 7);
+	double expected_values[] = {-2, -4, -6, 0};
 
-	float *v1 = sub_tuple(p1, p2);
+	double *v1 = sub_tuple(p1, p2);
 	for (int i = 0; i < 4; i++) {
 		cr_assert_eq(assert_float(v1[i],expected_values[i]), 1);
 	}
@@ -73,11 +73,11 @@ Test(sub_points, expected_vector) {
 
 Test(sub_vector_from_point, expected_sub_of_vectors_a_and_b)
 {
-	float *p1 = point(3, 2, 1);
-	float *v2 = vector(5, 6, 7);
-	float expected_values[] = {-2, -4, -6, 1};
+	double *p1 = point(3, 2, 1);
+	double *v2 = vector(5, 6, 7);
+	double expected_values[] = {-2, -4, -6, 1};
 
-	float *sub = sub_tuple(p1, v2);
+	double *sub = sub_tuple(p1, v2);
 	for (int i = 0; i < 4; i++) {
 		cr_assert_eq(assert_float(sub[i],expected_values[i]), 1);
 	}
@@ -87,8 +87,8 @@ Test(sub_vector_from_point, expected_sub_of_vectors_a_and_b)
 }
 
 Test(negate_vectors, expected_oposite_vector) {
-	float *v = vector(3, 2, 1);
-	float *nv = negate_tuple(v);
+	double *v = vector(3, 2, 1);
+	double *nv = negate_tuple(v);
 	for (int i = 0; i < 4; i++) {
 		cr_assert_eq(assert_float(nv[i], 0 - v[i]), 1);
 	}
@@ -97,9 +97,9 @@ Test(negate_vectors, expected_oposite_vector) {
 }
 
 Test(scalar_multiplication, one_vector_by_2) {
-	float *t1 = vector(2, 1, 4);
-	float *result = scalar_multiplication(t1, 2);
-	float expected_values[] = {4, 2, 8, 0};
+	double *t1 = vector(2, 1, 4);
+	double *result = scalar_multiplication(t1, 2);
+	double expected_values[] = {4, 2, 8, 0};
 	for (int i = 0; i < 4; i++) {
 		cr_assert_eq(assert_float(result[i], expected_values[i]), 1);
 	}
@@ -108,9 +108,9 @@ Test(scalar_multiplication, one_vector_by_2) {
 }
 
 Test(scalar_multiplication, one_vector_by_half) {
-	float *t1 = vector(2, 1, 4);
-	float *result = scalar_multiplication(t1, 0.5);
-	float expected_values[] = {1, 0.5, 2.0, 0};
+	double *t1 = vector(2, 1, 4);
+	double *result = scalar_multiplication(t1, 0.5);
+	double expected_values[] = {1, 0.5, 2.0, 0};
 	for (int i = 0; i < 4; i++) {
 		cr_assert_eq(assert_float(result[i], expected_values[i]), 1);
 	}
@@ -119,16 +119,16 @@ Test(scalar_multiplication, one_vector_by_half) {
 }
 
 Test(magnitude, expected_values_magnitude_1) {
-	float *v = vector(1, 0, 0);
-	float result = magnitude(v);
+	double *v = vector(1, 0, 0);
+	double result = magnitude(v);
 	cr_assert_eq(assert_float(result, 1), 1);
 	free((void*)v);
 }
 
 Test(magnitude, expected_value_magnitude_sqrt_of_14) {
-	float *v = vector(1, 2, 3);
-	float result = magnitude(v);
-	float expected_value = sqrtf(14);
+	double *v = vector(1, 2, 3);
+	double result = magnitude(v);
+	double expected_value = sqrtf(14);
 	cr_assert_eq(assert_float(result, expected_value), 1);
 	free((void*)v);
 }
@@ -151,8 +151,8 @@ Test(normalize, expected_vector_normalized_2)
 {
 	t_vector v = vector(1, 2, 3);
 	t_vector nv = normalize(v);
-	float	expected_value[] = {1.0, 2.0, 3.0, 0};
-	float	mag_expcted_value = sqrtf(14);
+	double	expected_value[] = {1.0, 2.0, 3.0, 0};
+	double	mag_expcted_value = sqrtf(14);
 	for (int i = 0; i < 4; i++) {
 		expected_value[i] *= (1 / mag_expcted_value);
 	}
@@ -166,7 +166,7 @@ Test(normalize, expected_vector_normalized_2)
 Test(dot_product, expected_0) {
 	t_vector v1 = vector(2, 0, 0);
 	t_vector v2 = vector(0, 2, 0);
-	float result = dot_product(v1, v2);
+	double result = dot_product(v1, v2);
 	cr_assert_eq(result, 0);
 	free((void*)v1);
 	free((void*)v2);
@@ -175,7 +175,7 @@ Test(dot_product, expected_0) {
 Test(dot_product, expected_12) {
 	t_vector v1 = vector(2, 2, 2);
 	t_vector v2 = vector(2, 2, 2);
-	float result = dot_product(v1, v2);
+	double result = dot_product(v1, v2);
 	cr_assert_eq(result, 12);
 	free((void*)v1);
 	free((void*)v2);
